moved secrets config to prod
